fix bug 2492545 (smbnetfs crash: Transport endpoint is not connected)
This is mix of two classical bugs: usage of pointer to freed data and
break of atomicity. We have the following sequence of operations

    smb_conn_close(fd->ctx, fd);
    samba_release_ctx(smb_conn_ctx_to_samba_ctx(fd->ctx));

In the first line we free the memory pointed by fd, so this memory may be
allocated and modified by the other thread. If this happened between first
and second lines, then the second operation will produce SIGSEGV as
fd->ctx will not point to correct context anymore.