add smb_timeout option
Thanks to Soci/Singular for his patch. It was a little bit modified througth.

Here is the description from Soci/Singular:

I've added "smb_timeout" configuration option, which is directly used with
smbc_setTimeout() function on connection init to setup the smbclient timeout.
Default is 20s, so it does not change anything by default. It does nothing
if smbclient is earlier than 3.2.0, but at least it still compiles.

Now that the timeouts can be controlled, I've noticed that the server processes
are killed on ETIMEDOUT and EHOSTUNREACH, which triggers the retry mechanism
for no good reason. Retries are ok if the processes are busy or killed for some
reason, but if smbclient times out on it's own (ETIMEDOUT) or the host is not
reachable at all then there's no point to exit (and retry). When the host is
back smbclient will continue without the need to exit the process.

ETIMEDOUT and EHOSTUNREACH are mapped to EIO.