add smb_timeout option (second attempt)
I throw out handling of ETIMEDOUT and EHOSTUNREACH from the original
patch as it already implemented.

Thanks to Soci/Singular for his patch, here is a part of original description

I've added "smb_timeout" configuration option, which is directly used with
smbc_setTimeout() function on connection init to setup the smbclient timeout.
Default is 20s, so it does not change anything by default. It does nothing
if smbclient is earlier than 3.2.0, but at least it still compiles.